# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Management API",
        "description": "API documentation for the Event Management System. This system allows managing yearly events with all their associated content including summaries, themes, programmes, speakers, sponsors, FAQs, media, galleries, and attendances.",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost",
            "description": "Production Server"
        }
    ],
    "paths": {
        "/api/events": {
            "get": {
                "tags": ["Events"],
                "summary": "Get all events",
                "description": "Returns a list of all events",
                "operationId": "getEvents",
                "parameters": [
                    {
                        "name": "published_only",
                        "in": "query",
                        "description": "Filter only published events",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {"type": "boolean", "example": true},
                                        "data": {
                                            "type": "array",
                                            "items": {"$ref": "#/components/schemas/Event"}
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["Events"],
                "summary": "Create a new event",
                "description": "Creates a new event",
                "operationId": "createEvent",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {"$ref": "#/components/schemas/EventRequest"}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Event created successfully",
                        "content": {
                            "application/json": {
                                "schema": {"$ref": "#/components/schemas/Event"}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/events/current": {
            "get": {
                "tags": ["Events"],
                "summary": "Get current year event",
                "description": "Returns the current year's published event with all related data",
                "operationId": "getCurrentEvent",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {"$ref": "#/components/schemas/Event"}
                            }
                        }
                    },
                    "404": {
                        "description": "Current event not found"
                    }
                }
            }
        },
        "/api/events/year/{year}": {
            "get": {
                "tags": ["Events"],
                "summary": "Get event by year",
                "description": "Returns an event for a specific year with all related data",
                "operationId": "getEventByYear",
                "parameters": [
                    {
                        "name": "year",
                        "in": "path",
                        "description": "Event year",
                        "required": true,
                        "schema": {"type": "integer", "example": 2026}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {"$ref": "#/components/schemas/Event"}
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            }
        },
        "/api/events/{id}": {
            "get": {
                "tags": ["Events"],
                "summary": "Get event by ID",
                "description": "Returns a single event with all related data",
                "operationId": "getEventById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event UUID",
                        "required": true,
                        "schema": {"type": "string", "format": "uuid"}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {"$ref": "#/components/schemas/Event"}
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            },
            "put": {
                "tags": ["Events"],
                "summary": "Update an event",
                "description": "Updates an existing event",
                "operationId": "updateEvent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event UUID",
                        "required": true,
                        "schema": {"type": "string", "format": "uuid"}
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {"$ref": "#/components/schemas/EventRequest"}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Event updated successfully"
                    },
                    "404": {
                        "description": "Event not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": ["Events"],
                "summary": "Delete an event",
                "description": "Deletes an event and all related data",
                "operationId": "deleteEvent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event UUID",
                        "required": true,
                        "schema": {"type": "string", "format": "uuid"}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event deleted successfully"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Event": {
                "type": "object",
                "properties": {
                    "id": {"type": "string", "format": "uuid", "example": "9b3d8f4e-7c1a-4d2b-9e5f-1a2b3c4d5e6f"},
                    "year": {"type": "integer", "example": 2026},
                    "title": {"type": "string", "example": "Annual Tech Conference 2026"},
                    "location": {"type": "string", "example": "Convention Center"},
                    "start_date": {"type": "string", "format": "date", "example": "2026-06-01"},
                    "end_date": {"type": "string", "format": "date", "example": "2026-06-03"},
                    "is_published": {"type": "boolean", "example": true},
                    "created_at": {"type": "string", "format": "date-time"},
                    "updated_at": {"type": "string", "format": "date-time"}
                }
            },
            "EventRequest": {
                "type": "object",
                "required": ["year", "title"],
                "properties": {
                    "year": {"type": "integer", "example": 2026},
                    "title": {"type": "string", "example": "Annual Tech Conference 2026"},
                    "location": {"type": "string", "example": "Convention Center"},
                    "start_date": {"type": "string", "format": "date", "example": "2026-06-01"},
                    "end_date": {"type": "string", "format": "date", "example": "2026-06-03"},
                    "is_published": {"type": "boolean", "example": true}
                }
            },
...
        }
    },
...
}